Update to reflect GPG key EOL & Slack
This addresses MycroftAI/enclosure-picroft#25 which I threw the answer on and can be closed.

This bit me, but with a totally different error message than in #25 and may put off users who aren't well versed in things like GPG/PGP/Apt, etc. Putting in the main Readme will hopefully allow users to fix the issue without too much frustration.

The only solution was buried in the slack channel, so I linked its existence at the bottom as well.