Initial check-in of the Stretch image
This new branch will host the Picroft components used in the Raspbian
Stretch Lite version.  This new version supports both the Raspberry Pi 3
and the 3B+, plus adds an interactive hardware setup wizard.

This also switches from being based on a Debian package install -- using
the package build for the Mark 1.  This is using a "Github setup"
approach.  This makes Picroft more developer-friendly, including the
ability to run the skill testing tools and mycroft-core unit tests.