Tweaks to auto_run script (#105)
* Remove double-run of start-mycroft.sh

The start-mycroft script was mistakenly being run twice on startup, including when SSHing in to the system.

Also switched from "source" to "bash".  There is not a good reason to run it in the current environment, and likely  unexpected side effects.

* Add software update after connecting to network

If a new version of software has been created, it should be pulled down as soon as possible before continuing with the setup_wizard.

* Add message for update process

The software update process can take a while, add a message warning users to be patient.

* Force all updates to happen at once

Moved call to dev_setup.sh to right after other updates.  This combines all the slow download steps into once chuck, limiting the number of times the user has to sit and wait.