Minor re-architecting of Picroft enclosures
This revamps how code is structured for enclosures.  Rather than repeating
code in several places, it now uses objects and inheritance.  The original
attempted to keep all the logic in a single file, but the GPIO objects
already added the need for a user to understand object oriented, so this
step seems reasonable.

The basics heirarchy is:
    Enclosure
        Picroft_Enclosure
            AIY_Enclosure
            Generic_Enclosure
            Matrix_Enclosure
            ReSpeaker_Enclosure

The Picroft_Enclosure contains most of the heavy lifting, with just a little
in the subclasses to tie to the special cases.

The GPIO_Button and GPIO_LED also implement must of the Raspberry Pi behavior.
The Button generates actions for Press, Release, Hold and Double Click.  The
GPIO_LED is a very simple LED which can turn on, off or flash automatically.

Also fixes the typo mentioned in Issue #125